🐞 fix: buld errors and warnings
